pull reading list update into separate file; use it in manual reading list addition
