+++ client/src/components/ReadingList/ManualReadingListAdder.tsx
@@ -1,13 +1,21 @@
 import { CheckCircleOutlined, CloseCircleOutlined, PlusOutlined } from '@ant-design/icons';
 import { Button, Input } from 'antd';
 import React, { useState } from 'react';
+import { useAppSelector } from '../../hooks/reduxHooks';
+import { blankPaper } from '../../templates';
+import { useUpdateReadingListFunc } from './hooks';
 
 export default function ManualReadingListAdder(): JSX.Element {
   const [isEditing, setIsEditing] = useState(false);
   const [title, setTitle] = useState('');
+  const readingList = useAppSelector((state) => state.readingList);
+
+  const { updateReadingListFunc } = useUpdateReadingListFunc();
 
   const submitItem = () => {
     setTitle('');
+    const newReadingList = readingList.concat({ ...blankPaper, title });
+    updateReadingListFunc(newReadingList);
   };
 
   return (
