+++ client/src/components/ReadingList/ReadingList-redux.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import axios from 'axios';
 import arrayMove from 'array-move';
 import { SortEndHandler } from 'react-sortable-hoc';
 import ReadingListContainer from './ReadingList-container';
@@ -8,23 +7,12 @@ import { blankNotes } from '../../templates';
 import { updateDraftId } from '../../slices/activeDraftSlice';
 import { setActiveReview } from '../../slices/activeReviewSlice';
 import { useAppDispatch, useAppSelector } from '../../hooks/reduxHooks';
-import { updateReadingList } from '../../slices/readingListSlice';
+import { useUpdateReadingListFunc } from './hooks';
 
 export default function ReadingListRedux(): JSX.Element {
   const dispatch = useAppDispatch();
   const readingList: Paper[] = useAppSelector((state) => state.readingList);
-  const { demoMode } = useAppSelector((state) => state.auth);
-
-  const updateReadingListFunc = async (newReadingList: Paper[]) => {
-    dispatch(updateReadingList(newReadingList));
-
-    // If we're in demoMode, don't attempt to update the server...
-    if (demoMode) {
-      return;
-    }
-    const { data } = await axios.put<Paper[]>('api/readingList', newReadingList);
-    dispatch(updateReadingList(data));
-  };
+  const { updateReadingListFunc } = useUpdateReadingListFunc();
 
   const handleEditClick = (paper: Paper) => {
     dispatch(updateDraftId(null));
